Only scan matching prefixes.
sled's scan method will _start_ at the correct portion of the keyspace, but
does not explicitly test whether you've run off the end. For these keys:

```
aa:bb
aa:cc
bb:aa
```
A scan over "aa" will not stop when it encounters "bb". We have to test for
this ourselves.